+++ config/disqus.php
@@ -3,5 +3,5 @@ return [
     'api_key'    => env('DISQUS_API_KEY', ''),
     'api_secret' => env('DISQUS_API_SECRET', ''),
     'domain'     => env('DISQUS_DOMAIN', ''), //<DISQUS_DOMAIN>.disqus.com/embed.js
-    'enabled'    => false,
+    'enabled'    => env('DISQUS_ENABLED', 'false'),
 ];