+++ tests/functional/ContentCest.php
@@ -24,7 +24,7 @@ class ContentCest {
 
         $I->canSee($translation->title);
         $I->canSee($translation->body);
-        $I->canSee($user->firstName . ' ' . $user->lastName);
+        $I->canSee($user->nickName);
         $I->canSee(date('d-m-Y', strtotime($content->createdAt)));
     }
 
@@ -186,7 +186,7 @@ class ContentCest {
         );
         $route              = '/' . $category->route->translations[0]['langCode'] . '/' . $category->route->translations[0]['url'];
 
-        $I->wantTo('see sticky content on the top of the list');
+        $I->wantTo('see promoted content on the top of the list');
         $I->amOnPage($route);
         $I->seeResponseCodeIs(200);