+++ tests/functional/LoginCest.php
@@ -19,6 +19,7 @@ class LoginCest {
         $I->wantTo('register a user');
 
         $I->amOnPage('/en/register');
+        $I->fillField('nickName', 'NickName');
         $I->fillField('firstName', 'John');
         $I->fillField('lastName', 'Doe');
         $I->fillField('email', 'example@example.com');
@@ -36,6 +37,7 @@ class LoginCest {
         $I->haveRecord(
             'Users',
             [
+                'nickName'  => 'NickName',
                 'firstName' => 'John',
                 'lastName'  => 'Doe',
                 'email'     => 'john@doe.com',
@@ -46,12 +48,14 @@ class LoginCest {
         );
 
         $I->amOnPage('/en/register');
+        $I->fillField('nickName', 'NickName');
         $I->fillField('firstName', 'John');
         $I->fillField('lastName', 'Doe');
         $I->fillField('email', 'john@doe.com');
         $I->fillField('password', 'password');
         $I->click('button[type=submit]');
-        $I->see('This email is already registered.');
+        $I->see('The email has already been taken.');
+        $I->see('The nick name has already been taken.');
         $I->dontSeeAuthentication();
     }
 
@@ -60,6 +64,7 @@ class LoginCest {
         $I->wantTo('prevent a spammer users registrations');
 
         $I->amOnPage('/en/register');
+        $I->fillField('nickName', 'NickName');
         $I->fillField('firstName', 'John');
         $I->fillField('lastName', 'Doe');
         $I->fillField('email', 'example@example.com');
@@ -137,5 +142,34 @@ class LoginCest {
         $I->login('john@doe.com', 'password');
         $I->logout();
         $I->dontSeeAuthentication();
+    }
+
+    public function canRemindPassword(FunctionalTester $I)
+    {
+        $I->wantTo('remind a password');
+        $I->amOnPage('/en/password/remind');
+        $I->see('Forgot password?');
+        $I->see('Enter the email address you used for creating your Account and we will send you instructions.');
+        $I->see('Return to sign in');
+    }
+
+    public function canNotRemindPasswordAsLoggedUser(FunctionalTester $I)
+    {
+        $I->wantTo('remind a password as logged user');
+        $I->haveRecord(
+            'Users',
+            [
+                'nickName'  => 'JohnDoe',
+                'firstName' => 'John',
+                'lastName'  => 'Doe',
+                'email'     => 'john@doe.com',
+                'password'  => bcrypt('password'),
+                'createdAt' => new DateTime(),
+                'updatedAt' => new DateTime(),
+            ]
+        );
+        $I->login('john@doe.com', 'password');
+        $I->amOnPage('/en/password/remind');
+        $I->see('JohnDoe');
     }
 }